GeglGtkView: Use floats instead of int for x/y translation

+++ gegl-gtk/gegl-gtk-view.h
@@ -57,11 +57,11 @@ GeglNode *gegl_gtk_view_get_node(GeglGtkView *self);
 void gegl_gtk_view_set_scale(GeglGtkView *self, float scale);
 float gegl_gtk_view_get_scale(GeglGtkView *self);
 
-void gegl_gtk_view_set_x(GeglGtkView *self, int x);
-int gegl_gtk_view_get_x(GeglGtkView *self);
+void gegl_gtk_view_set_x(GeglGtkView *self, float x);
+float gegl_gtk_view_get_x(GeglGtkView *self);
 
-void gegl_gtk_view_set_y(GeglGtkView *self, int y);
-int gegl_gtk_view_get_y(GeglGtkView *self);
+void gegl_gtk_view_set_y(GeglGtkView *self, float y);
+float gegl_gtk_view_get_y(GeglGtkView *self);
 
 G_END_DECLS
